backend/requirements.txt
fastapi
uvicorn
opencv-python
torch
torchvision
transformers
faiss-cpu
pillow
numpy
pymongo
imagehash
requests
python-multipart
scikit-learn
